pyproject.toml
[tool.poetry]
name = "drf-project-zzz"
version = "0.1.0"
description = ""
authors = ["PinderKingui1488 <ananimuszahar@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_project"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
serializer = "^0.2.1"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
